Fixed reference count underflow warning
The warning was caused when releasing the keychain. I think this is a
bug with SecKeychainOpen not actually retaining the keychain when in
Garbage Collection mode, but I'm not sure.

Fixed the problem by only releasing the keychain for non-Garbage
collected environments.

The warning appeared in the console log and I fixed this problem in
case this might be a reason for rejection from the App Store.