Adding in-app purchases support (fixes issue #17)
This patch includes support for InApp Purchases data on OSX 10.7 Lion,
while not breaking compatibility with 10.6.
The patch is limited to *extract* data from InApp Receipts: it DOES NOT
perform any validation on it! We shall discuss a way to validate InApp
Purchases (following the steps recommended by Apple) this with original
developers too.
The patch has been tested, though not intensively. It surely needs more
testing, and it needs to be tested with multiple items purchased
(currently I tested it only with 1 item purchased) and with big
quantities.